Refatora `include/` e reorganiza arquivos
