+++ include/core/common.h
