+++ include/core/state_manager.h
